Month 14/step 2

+++ src/controllers/cards.ts
@@ -1,18 +1,18 @@
-import { Request, Response } from 'express';
+import {NextFunction, Request, Response} from 'express';
 import Card from '../models/cards';
 import { IRequest } from '../types/Request';
 import {
   STATUS_OK,
   STATUS_BAD_REQUEST,
   STATUS_NOT_FOUND,
-  STATUS_SERVER_ERROR,
+  STATUS_SERVER_ERROR, STATUS_FORBIDDEN,
 } from '../constants/statusCodes';
 
-export const createCard = (req: IRequest, res: Response) => {
+export const createCard = (req: IRequest, res: Response, next: NextFunction) => {
   const { name, link } = req.body;
   const owner = req.user?._id;
   if (!name || !link) {
-    return res.status(STATUS_BAD_REQUEST).json({ message: 'Переданы некорректные данные при создании карточки' });
+    return next({ status: STATUS_BAD_REQUEST, message: 'Переданы некорректные данные при создании карточки, отсутствует имя или ссылка' });
   }
   return Card.create({
     name,
@@ -22,22 +22,22 @@ export const createCard = (req: IRequest, res: Response) => {
     .then((card) => res.status(STATUS_OK).json({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res.status(STATUS_BAD_REQUEST).json({ message: 'Переданы некорректные данные при создании карточки' });
+        return next({ status: STATUS_BAD_REQUEST, message: 'Переданы некорректные данные при создании карточки' });
       }
-      return res.status(STATUS_SERVER_ERROR).json({ message: 'Произошла ошибка' });
+      return next({ status: STATUS_SERVER_ERROR, message: 'Произошла ошибка' });
     });
 };
 
-export const getCards = (req: Request, res: Response) => Card.find({})
+export const getCards = (req: Request, res: Response, next: NextFunction) => Card.find({})
   .then((cards) => res.status(STATUS_OK).json({ data: cards }))
-  .catch(() => res.status(STATUS_SERVER_ERROR).json({ message: 'Произошла ошибка' }));
+    .catch(() => next({ status: STATUS_SERVER_ERROR, message: 'Произошла ошибка' }));
 
-export const likeCard = (req: IRequest, res: Response) => {
+export const likeCard = (req: IRequest, res: Response, next: NextFunction) => {
   const { cardId } = req.params;
   const userId = req.user?._id;
 
   if (!userId) {
-    return res.status(STATUS_BAD_REQUEST).json({ message: 'Переданы некорректные данные для постановки лайка' });
+    return next({ status: STATUS_BAD_REQUEST, message: 'Переданы некорректные данные для постановки лайка' });
   }
 
   Card.findByIdAndUpdate(
@@ -47,24 +47,24 @@ export const likeCard = (req: IRequest, res: Response) => {
   )
     .then((updatedCard) => {
       if (!updatedCard) {
-        return res.status(STATUS_NOT_FOUND).json({ message: 'Передан несуществующий _id карточки' });
+        return next({ status: STATUS_NOT_FOUND, message: 'Передан несуществующий _id карточки' });
       }
       return res.status(STATUS_OK).json({ data: updatedCard });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        return res.status(STATUS_BAD_REQUEST).json({ message: 'Передан некорректный _id карточки' });
+        return next({ status: STATUS_BAD_REQUEST, message: 'Передан некорректный _id карточки' });
       }
-      return res.status(STATUS_SERVER_ERROR).json({ message: 'Произошла ошибка' });
+      return next({ status: STATUS_SERVER_ERROR, message: 'Произошла ошибка' });
     });
 };
 
-export const dislikeCard = (req: IRequest, res: Response) => {
+export const dislikeCard = (req: IRequest, res: Response, next: NextFunction) => {
   const { cardId } = req.params;
   const userId = req.user?._id;
 
   if (!userId) {
-    return res.status(STATUS_BAD_REQUEST).json({ message: 'Переданы некорректные данные для снятия лайка' });
+    return next({ status: STATUS_BAD_REQUEST, message: 'Переданы некорректные данные для снятия лайка' });
   }
 
   Card.findByIdAndUpdate(
@@ -74,32 +74,50 @@ export const dislikeCard = (req: IRequest, res: Response) => {
   )
     .then((updatedCard) => {
       if (!updatedCard) {
-        return res.status(STATUS_NOT_FOUND).json({ message: 'Передан несуществующий _id карточки' });
+        return next({ status: STATUS_NOT_FOUND, message: 'Передан несуществующий _id карточки' });
       }
       return res.status(STATUS_OK).json({ data: updatedCard });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        return res.status(STATUS_BAD_REQUEST).json({ message: 'Передан некорректный _id карточки' });
+        return next({ status: STATUS_BAD_REQUEST, message: 'Передан некорректный _id карточки' });
       }
-      return res.status(STATUS_SERVER_ERROR).json({ message: 'Произошла ошибка' });
+      return next({ status: STATUS_SERVER_ERROR, message: 'Произошла ошибка' });
     });
 };
 
-export const deleteCardById = (req: Request, res: Response) => {
+export const deleteCardById = (req: IRequest, res: Response, next: NextFunction) => {
   const { id } = req.params;
+  const userId = req.user?._id;
 
-  Card.findByIdAndRemove(id)
-    .then((deletedCard) => {
-      if (!deletedCard) {
-        return res.status(STATUS_NOT_FOUND).json({ message: 'Карточка с указанным _id не найдена' });
+  Card.findById(id)
+    .then((card) => {
+      if (!card) {
+        return next({ status: STATUS_NOT_FOUND, message: 'Карточка с указанным _id не найдена' });
       }
-      return res.status(STATUS_OK).json({ data: deletedCard });
+
+      if (card.owner.toString() !== userId) {
+        return next({ status: STATUS_FORBIDDEN, message: 'У вас нет прав на удаление этой карточки' });
+      }
+
+      Card.findByIdAndRemove(id)
+        .then((deletedCard) => {
+          if (!deletedCard) {
+            return next({ status: STATUS_NOT_FOUND, message: 'Карточка с указанным _id не найдена' });
+          }
+          return res.status(STATUS_OK).json({ data: deletedCard });
+        })
+        .catch((err: any) => {
+          if (err.name === 'CastError') {
+            return next({ status: STATUS_BAD_REQUEST, message: 'Передан некорректный _id карточки' });
+          }
+          return next({ status: STATUS_SERVER_ERROR, message: 'Произошла ошибка' });
+        });
     })
-    .catch((err) => {
+    .catch((err: any) => {
       if (err.name === 'CastError') {
-        return res.status(STATUS_BAD_REQUEST).json({ message: 'Передан некорректный _id карточки' });
+        return next({ status: STATUS_BAD_REQUEST, message: 'Передан некорректный _id карточки' });
       }
-      return res.status(STATUS_SERVER_ERROR).json({ message: 'Произошла ошибка' });
+      return next({ status: STATUS_SERVER_ERROR, message: 'Произошла ошибка' });
     });
 };
